moved visitors to org.palladiosimulator.solver.core
