+++ bundles/org.palladiosimulator.solver/META-INF/MANIFEST.MF
@@ -43,7 +43,6 @@ Export-Package: org.palladiosimulator.solver,
  org.palladiosimulator.solver.transformations,
  org.palladiosimulator.solver.transformations.pcm2lqn,
  org.palladiosimulator.solver.transformations.pcm2regex,
- org.palladiosimulator.solver.visitors,
  org.palladiosimulator.solver.visualisation
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-Vendor: sdq.ipd.uka.de
